Prepared product for pic migration

+++ sephora-backend/Application/Dtos/Cart/CartDTO.cs
@@ -7,8 +7,8 @@ public class CartDto
     
     // public string UserId { get; set; } = String.Empty;
     
-    public int ProductId { get; set; }
-    public int ProductPieceId { get; set; }
+    public long ProductId { get; set; }
+    public long ProductPieceId { get; set; }
     public string ProductName { get; set; } = String.Empty;
     public string? ProductDescription { get; set; }
     public string ProductImage { get; set; } = String.Empty;
